reworked user deletion method

+++ src/main/java/com/agriguardian/service/AppUserService.java
@@ -159,7 +159,11 @@ public class AppUserService {
 
         try {
             List<AppUser> created = appUserRelationsRepository.findByControllerAndRelation(userRepo.findByUsername(username).get(), Relation.created).stream().map(appUserRelations -> appUserRelations.getUserFollower()).collect(Collectors.toList());
-            created.stream().forEach(appUser -> userRepo.delete(appUser));
+
+            if(created.size()!=0) {
+                created.stream().forEach(appUser -> userRepo.delete(appUser));
+
+            }
             userRepo.deleteByUsername(username);
         } catch (Exception e) {
             log.error("[deleteUser] failed to delete a user {}; rsn: {}", username, e.getMessage());
